norm, edit lexer, edit ft_strerror

+++ src/utils/key_exists.c
@@ -7,9 +7,9 @@
 int	key_exists(char	*key, char **envp)
 {
 	int	i;
-	int key_len;
+	int	key_len;
 
-	key_len = ft_strlen(key);
+	key_len = (int)ft_strlen(key);
 	if (envp)
 	{
 		i = -1;
